NOJIRA-Billing-ledger-refactoring
Fix review findings and add comprehensive test coverage for billing deduction logic.

- bin-api-manager: Fix float32 truncation in balance endpoints using math.Round
- bin-api-manager: Add missing return after AbortWithStatus(400) in balance endpoints
- bin-billing-manager: Extract CalculateTokenCreditDeduction as pure testable function
- bin-billing-manager: Fix negative token balance bug causing incorrect deductions
- bin-billing-manager: Add 25 edge case tests for token-credit deduction logic
- bin-billing-manager: Add 8 edge cases to CalculateBillableUnits tests
- bin-billing-manager: Change PlanTokenMap from int to int64 for type consistency
- bin-dbscheme-manager: Fix CEIL(NULL) in migration with COALESCE

+++ bin-billing-manager/cmd/billing-control/main.go
@@ -615,7 +615,7 @@ func runTopUpRun(cmd *cobra.Command, args []string) error {
 			continue
 		}
 
-		if err := db.AccountTopUpTokens(context.Background(), a.ID, a.CustomerID, int64(tokenAmount), string(a.PlanType)); err != nil {
+		if err := db.AccountTopUpTokens(context.Background(), a.ID, a.CustomerID, tokenAmount, string(a.PlanType)); err != nil {
 			fmt.Printf("ERROR: account_id=%s err=%v\n", a.ID, err)
 			continue
 		}
